requirements.txt
matplotlib==3.4.3
numpy==1.26.3
opencv-python-headless==4.5.3.56
Pillow==8.4.0
protobuf==3.18.0
rembg==0.1.1
requests==2.26.0
scikit-learn==1.4.0
scipy==1.12.0
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.9
faiss-cpu==1.7.4
